v0.0.771.0 perf: adaptive chunk size for chunked PUT (#984)
* perf: adaptive chunk size for chunked PUT (v0.0.771.0)

A 10MiB object fit in one 16MiB chunk = single segment = the 8-worker
SegmentWriter using only 1 worker with no read/EC/write pipeline. The GCP
size x chunk sweep showed this is the PUT bottleneck (CPU stayed ~58%; the loss
is serial latency, not utilization):
  - 10MiB at 2 segments: 391 -> 446 MB/s (+13.7%, 0.95x of MinIO)
  - but small objects LOSE when split (1MiB fell 288 -> 157 MB/s; per-segment
    EC/shard-write/staging-promote overhead dominates), and over-splitting
    (5 segments) was worse than 2.

adaptiveChunkSize(objectSize) = clamp(objectSize/2, 5MiB floor, DefaultChunkSize
cap): small objects stay single, mid-size split into 2, large cap and split
naturally. Multipart-complete (worker=1) and unknown-size keep the default.
The chunk-size helpers also return int64 (byte count) so callers drop their
int64(...) casts.

* docs: refresh cluster benchmark with adaptive-chunk result

Update GCP 4-node cluster numbers in README and benchmarks.md to the adaptive
chunk-size run (PUT 446 MiB/s = 0.95x MinIO, GET 1998 = 1.02x, profiling off).
Collapse the benchmarks.md cluster table to the same 4-column throughput shape
as the single-node table; p50/p99 latency detail moves into interpretation prose
since warp-results.tsv carries throughput only.

+++ internal/cluster/backend.go
@@ -229,7 +229,7 @@ type DistributedBackend struct {
 	qms *QuorumMetaStore
 
 	// chunkedPutChunkSize is a test seam; zero keeps the production default.
-	chunkedPutChunkSize int
+	chunkedPutChunkSize int64
 
 	// bypassBucketCheck skips the HeadBucket pre-check in PutObject. Set by
 	// GroupBackend: bucket existence is guaranteed by the router (design doc
